core/firmware_editor.py
"""
Firmware Editor Module
Provides tools for editing firmware partitions
"""

import os
import logging
import shutil
from pathlib import Path
from typing import Optional, Callable, Dict, List
from PIL import Image
import struct


class FirmwareEditor:
    """Edit and modify Android firmware partitions"""
    
    def __init__(self, config):
        self.config = config
        self.logger = logging.getLogger(__name__)
        self.workspace = None
        self.partitions = {}
        self.changes_log = []
        
    def create_workspace(self, base_dir: Path) -> Path:
        """Create working directory for firmware modifications"""
        self.workspace = base_dir / "firmware_workspace"
        self.workspace.mkdir(parents=True, exist_ok=True)
        
        # Create subdirectories
        (self.workspace / "original").mkdir(exist_ok=True)
        (self.workspace / "modified").mkdir(exist_ok=True)
        (self.workspace / "output").mkdir(exist_ok=True)
        
        self.logger.info(f"Workspace created: {self.workspace}")
        return self.workspace
    
    def load_partition(self, partition_name: str, file_path: Path) -> bool:
        """Load a partition file for editing"""
        try:
            if not file_path.exists():
                raise FileNotFoundError(f"Partition file not found: {file_path}")
            
            # Copy to workspace
            workspace_path = self.workspace / "original" / file_path.name
            shutil.copy(file_path, workspace_path)
            
            self.partitions[partition_name] = {
                'original': workspace_path,
                'modified': self.workspace / "modified" / file_path.name,
                'size': os.path.getsize(workspace_path)
            }
            
            # Copy to modified as starting point
            shutil.copy(workspace_path, self.partitions[partition_name]['modified'])
            
            self.logger.info(f"Loaded partition: {partition_name}")
            return True
            
        except Exception as e:
            self.logger.error(f"Failed to load partition {partition_name}: {e}")
            return False
    
    def modify_boot_logo(self, logo_image_path: Path) -> bool:
        """
        Replace boot logo in logo partition
        
        Args:
            logo_image_path: Path to new logo image (should be PNG/BMP)
        
        Returns:
            True if successful
        """
        if 'logo' not in self.partitions:
            self.logger.error("Logo partition not loaded")
            return False
        
        try:
            self.logger.info("Modifying boot logo...")
            
            # Open and resize logo image
            img = Image.open(logo_image_path)
            target_size = self.config.get('preview.screen_resolution', [1080, 1920])
            img = img.resize((target_size[0], target_size[1]), Image.Resampling.LANCZOS)
            
            # Convert to format expected by MTK boot logo
            # MTK typically uses RGB565 or RGB888 format
            img_rgb = img.convert('RGB')
            
            # Create logo partition data
            logo_data = self._create_mtk_logo_data(img_rgb)
            
            # Write to modified partition
            modified_path = self.partitions['logo']['modified']
            with open(modified_path, 'wb') as f:
                f.write(logo_data)
            
            self.changes_log.append({
                'partition': 'logo',
                'action': 'replace_boot_logo',
                'source': str(logo_image_path)
            })
            
            self.logger.info("Boot logo modified successfully")
            return True
            
        except Exception as e:
            self.logger.error(f"Failed to modify boot logo: {e}")
            return False
    
    def _create_mtk_logo_data(self, img: Image.Image) -> bytes:
        """Convert PIL image to MTK logo partition format"""
        # MTK Logo format header
        # This is a simplified version - actual format may vary by device
        
        width, height = img.size
        
        # Header: magic + width + height
        header = struct.pack('<4sII', b'LOGO', width, height)
        
        # Convert to RGB565 (common format for MTK logos)
        img_data = bytearray()
        pixels = img.load()
        
        for y in range(height):
            for x in range(width):
                r, g, b = pixels[x, y]
                # Convert 8-bit RGB to RGB565
                rgb565 = ((r >> 3) << 11) | ((g >> 2) << 5) | (b >> 3)
                img_data.extend(struct.pack('<H', rgb565))
        
        return header + bytes(img_data)
    
    def modify_build_prop(self, properties: Dict[str, str]) -> bool:
        """
        Modify build.prop in system partition
        
        Args:
            properties: Dictionary of property key-value pairs to modify
        
        Returns:
            True if successful
        """
        if 'system' not in self.partitions:
            self.logger.error("System partition not loaded")
            return False
        
        try:
            self.logger.info("Modifying build.prop...")
            
            # TODO: Implement system.img mounting and build.prop editing
            # This requires ext4 filesystem manipulation
            
            self.changes_log.append({
                'partition': 'system',
                'action': 'modify_build_prop',
                'properties': properties
            })
            
            return True
            
        except Exception as e:
            self.logger.error(f"Failed to modify build.prop: {e}")
            return False
    
    def inject_file(self, partition_name: str, 
                   source_file: Path, 
                   target_path: str) -> bool:
        """
        Inject a file into a partition
        
        Args:
            partition_name: Name of partition (e.g., 'system', 'vendor')
            source_file: File to inject
            target_path: Path within partition where file should be placed
        
        Returns:
            True if successful
        """
        if partition_name not in self.partitions:
            self.logger.error(f"Partition {partition_name} not loaded")
            return False
        
        try:
            self.logger.info(f"Injecting {source_file} into {partition_name}:{target_path}")
            
            # TODO: Implement filesystem mounting and file injection
            
            self.changes_log.append({
                'partition': partition_name,
                'action': 'inject_file',
                'source': str(source_file),
                'target': target_path
            })
            
            return True
            
        except Exception as e:
            self.logger.error(f"Failed to inject file: {e}")
            return False
    
    def get_changes_summary(self) -> List[Dict]:
        """Get list of all modifications made"""
        return self.changes_log.copy()
    
    def export_modified_firmware(self, output_dir: Path) -> Dict[str, Path]:
        """
        Export all modified partitions
        
        Args:
            output_dir: Directory to export modified partitions
        
        Returns:
            Dictionary mapping partition names to their exported paths
        """
        output_dir.mkdir(parents=True, exist_ok=True)
        exported = {}
        
        for name, info in self.partitions.items():
            modified_path = info['modified']
            output_path = output_dir / modified_path.name
            
            shutil.copy(modified_path, output_path)
            exported[name] = output_path
            
            self.logger.info(f"Exported {name} to {output_path}")
        
        return exported
    
    def verify_partition(self, partition_name: str) -> Dict:
        """
        Verify partition integrity
        
        Returns:
            Dictionary with verification results
        """
        if partition_name not in self.partitions:
            return {'valid': False, 'error': 'Partition not loaded'}
        
        info = self.partitions[partition_name]
        modified_path = info['modified']
        
        if not modified_path.exists():
            return {'valid': False, 'error': 'Modified partition file missing'}
        
        # Check file size
        modified_size = os.path.getsize(modified_path)
        original_size = info['size']
        
        result = {
            'valid': True,
            'original_size': original_size,
            'modified_size': modified_size,
            'size_changed': modified_size != original_size
        }
        
        return result
